+++ android/app/src/main/java/com/borasarang/jupjup/JupJupApplication.kt
@@ -1,10 +1,8 @@
 package com.borasarang.jupjup
 
 import android.app.Application
-import com.borasarang.communityjupjup.CommunityJupJupRuntime
 import com.borasarang.macjupjup.MacJupJupRuntime
 import com.borasarang.planjupjup.PlanJupJupRuntime
-import com.borasarang.promptjournaljupjup.PromptJournalRuntime
 
 /**
  * 줍줍 시리즈 통합 앱.
@@ -17,7 +15,5 @@ class JupJupApplication : Application() {
         super.onCreate()
         MacJupJupRuntime.initialize(this)
         PlanJupJupRuntime.initialize(this)
-        PromptJournalRuntime.initialize(this)
-        CommunityJupJupRuntime.initialize(this)
     }
 }